requirements.txt
beautifulsoup4==4.12.2
selenium==4.11.2
pytz==2021.3
httpx==0.24.1
python-telegram-bot==20.5
